Add remote pregen tooling and tail-50k chunk manifests
Add scripts/pregen_remote.sh (SLURM array pregen, 8-way strided shards)
and remote_pregen/ chunk manifests + rsync file-lists covering the last
~48k uncached samples, split into <=90GB video chunks. Manifests carry
only video+prompt (the gen-cache key) to stay small. Includes the
accumulated training/pregen code changes (multi-variant trainer,
resumable sampler, world-size-agnostic checkpointing, qformer 1024).

+++ vflash/models/layers.py
@@ -38,10 +38,11 @@ class RotaryEmbedding(nn.Module):
 
 
 def apply_rope(x, cos, sin):
-    # x [B,H,S,hd]; cos/sin [B,S,hd]
+    # x [B,H,S,hd]; cos/sin [B,S,hd]. RoPE math in fp32, result cast back to x's
+    # dtype so q/k match v (flex_attention/SDPA require identical q,k,v dtypes).
     cos = cos.unsqueeze(1)
     sin = sin.unsqueeze(1)
-    return x * cos + rotate_half(x) * sin
+    return (x * cos + rotate_half(x) * sin).to(x.dtype)
 
 
 class MLP(nn.Module):
